Treat .url files as shortcuts
The Epic store creates .url shortcut files to all the games it installs,
which would ideally be tidied up.

It seems like Shellify lib isn't able to read the .url files, so it's
not currently possible to support the .url file type for "App Shortcuts"
filtering. If you want .url files to be deleted, you must use the "All
Shortcuts" setting.